fix format type

+++ components/somfy/Somfy.cpp
@@ -40,7 +40,7 @@ void SomfyComponent::send_command_internal(Command command) {
 
 void SomfyComponent::send_command(Command command) {
     const uint16_t rollingCode = this->storage_->nextCode();
-    ESP_LOGD(TAG, "Sending %dx command: 0x%x for button addr: 0x%" PRIx32 " with rolling code: 0x%04X", this->repeat_, command, this->remote_address_, rollingCode);
+    ESP_LOGD(TAG, "Sending %dx command: 0x%x for button addr: 0x%" PRIx32 " with rolling code: 0x%04X", this->repeat_, static_cast<unsigned int>(command), this->remote_address_, rollingCode);
     uint8_t frame[7];
     build_frame(frame, command, rollingCode);
     remote_base::RawTimings t;
